🎨(core) rename utils module to core
The name `core` seems to be better suited for the module containing
the main part of this library.
