+++ src/multicons/core.py
